# src/main/resources/swagger.yaml
openapi: 3.0.2
info:
  title: generic-project
  version: v1
tags:
  - name: auth
    description: Endpoints de autenticación
security:
  - bearerAuth: []
paths:
  /users/test:
    get:
      summary: endpoint publico test
      operationId: getTest
      responses:
        '200':
          $ref: '#/components/responses/ReponseTest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /login:
    post:
      summary: Login de un user
      operationId: login
      requestBody:
        $ref: '#/components/requestBodies/AuthenticationRequestBody'
      responses:
        '200':
          $ref: '#/components/responses/AuthenticationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - auth
  /register:
    post:
      summary: Register de un user
      operationId: register
      requestBody:
        $ref: '#/components/requestBodies/RegisterRequestBody'
      responses:
        '200':
          $ref: '#/components/responses/AuthenticationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - auth
components:
  requestBodies:
    AuthenticationRequestBody:
      description: Login de un user
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationRequestDTO'
          example:
            email: "usuario@ejemplo.com"
            password: "123456"
    RegisterRequestBody:
      description: Registrar un user
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegisterRequestDTO'

  parameters:
    genericId:
      name: genericId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/genericId'
  responses:
    ReponseTest:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TestDTO'
    AuthenticationResponse:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationDTO'
    NoContent:
      description: No Content
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 400
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Unauthorized:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 401
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Forbidden:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 403
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    NotFound:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    UnexpectedError:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        email:
          type: string
          example: user@user.com
        firstname:
          type: string
          example: Luis
        lastname:
          type: string
          example: Perez
        isEmailActive:
          type: boolean
          example: false
        role:
          type: string
          example: USER
        tier:
          type: string
          example: FREE
    AuthenticationRequestDTO:
      type: object
      properties:
        email:
          type: string
          example: friasgilivan@gmail.com
        password:
          type: string
          example: 123456
    TestDTO:
      type: object
      properties:
        version:
          type: string
          example: 0.0.2
    RegisterRequestDTO:
      type: object
      properties:
        email:
          type: string
          example: friasgilivan@gmail.com
        password:
          type: string
          example: 123456
        firstname:
          type: string
          example: 123456
        lastname:
          type: string
          example: 123456
        role:
          $ref: '#/components/schemas/RoleUser'
    AuthenticationDTO:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9.eyJmaXJzdG5hbW...
    RoleUser:
      type: string
      enum:
        - USER
        - ADMIN
      description: Role de los users de la app
    Problem:
      type: object
      properties:
        error:
          type: string
    genericId:
      title: example
      description: example
      type: integer
      format: int64
      minimum: 1
      maximum: 999999999999
      example: 1

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT