# target/generated-sources/openapi/src/main/java/com/mycompany/bookme/api/RegisterApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mycompany.bookme.api;

import com.mycompany.bookme.model.AuthenticationDTO;
import com.mycompany.bookme.model.Problem;
import com.mycompany.bookme.model.RegisterRequestDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-10-30T10:31:37.615691+01:00[Europe/Madrid]")
@Validated
@Tag(name = "auth", description = "Endpoints de autenticación")
public interface RegisterApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /register : Register de un user
     *
     * @param registerRequestDTO Registrar un user (required)
     * @return OK. (status code 200)
     *         or Bad Request (status code 400)
     *         or Bad Request (status code 401)
     *         or Bad Request (status code 403)
     *         or Bad Request (status code 404)
     *         or Bad Request (status code 500)
     */
    @Operation(
        operationId = "register",
        summary = "Register de un user",
        tags = { "auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthenticationDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Problem.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Bad Request", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Problem.class)))
            }),
            @ApiResponse(responseCode = "403", description = "Bad Request", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Problem.class)))
            }),
            @ApiResponse(responseCode = "404", description = "Bad Request", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Problem.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Bad Request", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Problem.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/register",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<AuthenticationDTO> register(
        @Parameter(name = "RegisterRequestDTO", description = "Registrar un user", required = true) @Valid @RequestBody RegisterRequestDTO registerRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"token\" : \"eyJhbGciOiJIUzI1NiJ9.eyJmaXJzdG5hbW...\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"error\" : \"error\" }, { \"error\" : \"error\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"error\" : \"error\" }, { \"error\" : \"error\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"error\" : \"error\" }, { \"error\" : \"error\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"error\" : \"error\" }, { \"error\" : \"error\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"error\" : \"error\" }, { \"error\" : \"error\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}